fix show image on server

+++ database/migrations/2021_09_22_073854_create_uploads_table.php
@@ -20,7 +20,7 @@ class CreateUploadsTable extends Migration
             $table->string('filename');
             $table->string('original_name');
             $table->string('url');
-            $table->float('size', 5, 1);
+            $table->unsignedInteger('size');
             $table->timestamps();
 
             $table->foreign('id_users')->references('id')->on('users')
